The State of the Exchange
This is it. This is the current state of the app.

It went through some growing pains last season, so a lot of the patches
were probably not applied with the greatest of care. Things are kind of
a mess.

Lots of new ideas to build on, but plenty to clean up.